handdetection/requirements.txt
flask==3.0.2
opencv-python==4.9.0.80
numpy==1.26.4
tensorflow==2.16.1
gTTS==2.5.1
google-generativeai==0.3.2 